# storage/api-docs/api-docs.yaml/api-docs.yaml
openapi: 3.0.3
info:
  title: 'Organizations API'
  description: 'API для работы с организациями, их видами деятельности и географическими данными'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:85/api'
    description: 'Локальный сервер'
  -
    url: 'http://localhost:85/api'
paths:
  '/organizations/building/{building}':
    get:
      tags:
        - Organizations
      summary: 'Список организаций в конкретном здании'
      operationId: 448081f42d453487d3982d67ef030b01
      parameters:
        -
          name: building
          in: path
          description: 'ID здания'
          required: true
          schema:
            type: integer
        -
          name: per_page
          in: query
          description: 'Количество записей на страницу'
          schema:
            type: integer
            default: 15
            maximum: 500
            minimum: 1
      responses:
        200:
          description: 'Успешный ответ'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        401:
          description: Неавторизован
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: 'Unauthorized: Invalid or missing API key'
                type: object
        422:
          description: 'Ошибка валидации'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          ApiKeyAuth: []
  '/organizations/activity/{activity}':
    get:
      tags:
        - Organizations
      summary: 'Список организаций по виду деятельности'
      operationId: ccb9e4f9f903c58edec3f905c10c4fed
      parameters:
        -
          name: activity
          in: path
          description: 'ID вида деятельности'
          required: true
          schema:
            type: integer
        -
          name: per_page
          in: query
          description: 'Количество записей на страницу'
          schema:
            type: integer
            default: 15
            maximum: 500
            minimum: 1
      responses:
        200:
          description: 'Успешный ответ'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        401:
          description: Неавторизован
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: 'Unauthorized: Invalid or missing API key'
                type: object
        422:
          description: 'Ошибка валидации'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          ApiKeyAuth: []
  /organizations/geo/rectangle:
    get:
      tags:
        - Organizations
      summary: 'Список организаций в заданной прямоугольной области'
      operationId: c4a658e0ffd6c897bba4131c014715a3
      parameters:
        -
          name: min_lat
          in: query
          description: 'Минимальная широта'
          required: true
          schema:
            type: number
            format: float
            maximum: 90
            minimum: -90
        -
          name: max_lat
          in: query
          description: 'Максимальная широта'
          required: true
          schema:
            type: number
            format: float
            maximum: 90
            minimum: -90
        -
          name: min_lon
          in: query
          description: 'Минимальная долгота'
          required: true
          schema:
            type: number
            format: float
            maximum: 180
            minimum: -180
        -
          name: max_lon
          in: query
          description: 'Максимальная долгота'
          required: true
          schema:
            type: number
            format: float
            maximum: 180
            minimum: -180
        -
          name: per_page
          in: query
          description: 'Количество записей на страницу'
          schema:
            type: integer
            default: 100
            maximum: 500
            minimum: 1
      responses:
        200:
          description: 'Успешный ответ'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        401:
          description: Неавторизован
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: 'Unauthorized: Invalid or missing API key'
                type: object
        422:
          description: 'Ошибка валидации'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          ApiKeyAuth: []
  /organizations/geo/radius:
    get:
      tags:
        - Organizations
      summary: 'Список организаций в заданном радиусе'
      operationId: 0ee00f7231a7ce3a3f248f2023d02cd3
      parameters:
        -
          name: lat
          in: query
          description: 'Широта центра области'
          required: true
          schema:
            type: number
            format: float
            maximum: 90
            minimum: -90
        -
          name: lon
          in: query
          description: 'Долгота центра области'
          required: true
          schema:
            type: number
            format: float
            maximum: 180
            minimum: -180
        -
          name: radius
          in: query
          description: 'Радиус в километрах'
          schema:
            type: number
            format: float
            default: 15
            maximum: 1000
            minimum: 0
        -
          name: per_page
          in: query
          description: 'Количество записей на страницу'
          schema:
            type: integer
            default: 50
            maximum: 500
            minimum: 1
      responses:
        200:
          description: 'Успешный ответ'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        401:
          description: Неавторизован
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: 'Unauthorized: Invalid or missing API key'
                type: object
        422:
          description: 'Ошибка валидации'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          ApiKeyAuth: []
  '/organizations/{id}':
    get:
      tags:
        - Organizations
      summary: 'Информация об организации по ID'
      operationId: f9f68d1d824e41ffd87f55c2d19e4a86
      parameters:
        -
          name: id
          in: path
          description: 'ID организации'
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Успешный ответ'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        401:
          description: Неавторизован
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: 'Unauthorized: Invalid or missing API key'
                type: object
        404:
          description: 'Организация не найдена'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'No query results for model [App\Models\Organization].'
                type: object
      security:
        -
          ApiKeyAuth: []
  /organizations/search:
    get:
      tags:
        - Organizations
      summary: 'Поиск организаций по названию'
      operationId: 610f093d6e00fbea78c1ec2e24f595ba
      parameters:
        -
          name: name
          in: query
          description: 'Название организации для поиска (частичное совпадение)'
          schema:
            type: string
            maxLength: 255
        -
          name: per_page
          in: query
          description: 'Количество записей на страницу'
          schema:
            type: integer
            default: 15
            maximum: 500
            minimum: 1
      responses:
        200:
          description: 'Успешный ответ'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        401:
          description: Неавторизован
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: 'Unauthorized: Invalid or missing API key'
                type: object
        422:
          description: 'Ошибка валидации'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          ApiKeyAuth: []
  '/organizations/activity-with-children/{activity}':
    get:
      tags:
        - Organizations
      summary: 'Поиск организаций по виду деятельности и его дочерним категориям'
      operationId: a800dc243a3b016e8979dd9979e2d148
      parameters:
        -
          name: activity
          in: path
          description: 'ID вида деятельности'
          required: true
          schema:
            type: integer
        -
          name: per_page
          in: query
          description: 'Количество записей на страницу'
          schema:
            type: integer
            default: 15
            maximum: 500
            minimum: 1
      responses:
        200:
          description: 'Успешный ответ'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        401:
          description: Неавторизован
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: 'Unauthorized: Invalid or missing API key'
                type: object
        422:
          description: 'Ошибка валидации'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          ApiKeyAuth: []
components:
  schemas:
    PaginatedResponse:
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Organization'
        meta:
          properties:
            current_page:
              type: integer
              example: 1
            per_page:
              type: integer
              example: 15
            total:
              type: integer
              example: 50
            last_page:
              type: integer
              example: 4
          type: object
      type: object
    ErrorResponse:
      properties:
        error:
          type: string
          example: 'Validation failed'
        messages:
          type: object
          example:
            per_page: 'Параметр per_page должен быть целым числом.'
            lat: 'Параметр lat обязателен.'
          additionalProperties: true
      type: object
    Activity:
      properties:
        id:
          type: integer
          example: 12
        name:
          type: string
          example: Аксессуары
        parent_id:
          type: integer
          example: 2
          nullable: true
        level:
          type: integer
          example: 2
        children:
          type: array
          items:
            $ref: '#/components/schemas/Activity'
        created_at:
          type: string
          format: date-time
          example: '2025-10-01T13:07:29.000000Z'
        updated_at:
          type: string
          format: date-time
          example: '2025-10-01T13:07:29.000000Z'
      type: object
    Building:
      properties:
        id:
          type: integer
          example: 21
        address:
          type: string
          example: 'г. Москва, ул. Тестовая, д. 1'
        coordinates:
          properties:
            latitude:
              type: number
              format: float
              example: 55.6645
            longitude:
              type: number
              format: float
              example: 37.6168
          type: object
        organizations_count:
          type: integer
          example: 14
        created_at:
          type: string
          format: date-time
          example: '2025-10-01T13:07:29.000000Z'
        updated_at:
          type: string
          format: date-time
          example: '2025-10-01T13:07:29.000000Z'
      type: object
    Organization:
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'Тестовая организация'
        building:
          $ref: '#/components/schemas/Building'
        phones:
          type: array
          items:
            type: string
            example: +7-900-871-25-72
        activities:
          type: array
          items:
            $ref: '#/components/schemas/Activity'
        distance:
          type: number
          format: float
          example: 4.7996408640902
          nullable: true
        created_at:
          type: string
          format: date-time
          example: '2025-10-01T13:07:30.000000Z'
        updated_at:
          type: string
          format: date-time
          example: '2025-10-01T13:07:30.000000Z'
      type: object
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      description: 'API ключ для авторизации'
      name: X-API-KEY
      in: header
tags:
  -
    name: Organizations
    description: 'Операции с организациями'
security:
  -
    ApiKeyAuth: []